this fixes 3 separate things reported by robl and phil.  this changelog lists phil's description and a description of the fix:
---
1) This one isn't too bad, but if I run this:
      ./pvfs2-client -p pvfs2-client-core
   instead of this:
      ./pvfs2-client -p ./pvfs2-client-core
  ... the output looks the same either way (ie, it appears successful).
In the former case, both the client and client-core actually exited,
because i specified the path to the core incorrectly.  It would be nice
if it printed something in this case to avoid confusion.
---

Ok, done...this is tricky because if the client-core is in the same dir as
the client program, a 'stat' will pass, but the execvp may fail if that dir
isn't in the system PATH.  Now, we print an error if the path is not absolute
and the client core exits.

---
2) I don't know if this happens everywhere or not, but I am running on a
box here where bad things happen if I get the hostname wrong on the
mount command, for example doing something like this:
        mount -t pvfs2 tcp://a43:0/pvfs2-fs /mnt/pvfs2
(note that the port is wrong).  Sorry I didn't catch this when you asked
me to try this stuff out the other day :(  At any rate, the mount command
segfaults, and I get this in dmesg:

Attempting PVFS2 Mount via host tcp://a43:0/pvfs2-fs
Got an unknown pvfs2 error code: -1073741967
pvfs2_fs_mount: got return value of -1073741967
Unable to handle kernel paging request at virtual address 733d4853
 printing eip:
 c016b062
 *pde = 00000000
---

This is due to an error code translation problem.  Previously, the kernel
code didn't translate all pvfs_errno codes to actual errno codes, so I've
re-worked the error code translation functions so that they can be shared
by the kernel and the user space code.

---
3) This one is kinda nasty, and I'm not sure what's going wrong yet.
The scenario is that I have successfully mounted, but then on the next
operation (a getattr), the communication fails causing the msgpair.sm to
call exit(1).  The pvfs2-client and client-core then exit.  I start them
back up, and nothing ever works again.  I get "Input/output error" any
time I try to access pvfs2 from then on, and I get "can't write
superblock" when I try to unmount.  I have to reboot the box to get back
to sanity.
---

This turned out to be a beast of a problem now that we have dynamic mounts on
the client side.  The issue is that on cancelled i/o, or on pvfs2-client-core
restart, there's *no* way the kernel can associate upcalls with the client-core
anymore because the dynamic mount information is lost forever as far as the
client-core is concerned.

The solution I chose for this is to have the pvfs2-client-core issue an ioctl
that causes the kernel to do mount upcalls for every pvfs2 file system it knows
about to avoid having to have the user manually issue another mount command
(or something goofy like that).  Other solutions that I considered and didn't
like are 1) having the pvfs2-client-core store mount info in a file somewhere
and have it remount those on startup, or 2) having the pvfs2-client and
pvfs2-client-core have some ipc going on so that the pvfs2-client can store
the mount info and somehow give it back to the pvfs2-client-core on restart.
both of those are shoddy at best, and do not seem robust at all (i.e.
stale file info will error out on stale remount attempts, ipc requires that
the pvfs2-client never dies along with the pvfs2-client-core, etc).

With the current solution, we have all the information that the kernel knows
about (and so it's more reliable than a file or something), and it's the most
natural way to hand the information up from the kernel (since we have to do
this via upcalls anyway).  Thus, the kernel code now keeps a list of pvfs2
superblocks and we store the info given to the sb at mount time so that it
can dynamically be remounted at any time.  The pvfs2-client-core now issues
an ioctl (in a thread so that it can still service the request) on startup
to 'remount' any file systems the kernel knows about.  while I'm calling this
a remount, it's basically just a mechanism for the kernel to tell the client
core enough information ot rebuild the dynamic mount tables in the system
interface.

because of this, the pvfs2-client-core program *requires* pthreads.  This is
completely independent of whether or not the PVFS2 sysint needs pthreads
at all.


SVN revision 3294