Adding support for get_config state variable persist_config_buffers.
When 0 the existing behavior is reproduced, the config buffers are free'd
when the get_config is completed.  When persist_config_buffers is 1,
fs_config_buf and server_config_buf are not free'd by the state machine and
must be free'd by the user.  This was added so that pvfs2-check-config can
compare each servers config files.  Checkpoint checkin.


SVN revision 3466