[pcarns]: force the st_nlink field to stay fixed at 1.  The find(1) program will get confused if we fix the field at 2 and quit too early.
"The "normal" semantic is for st_nlink to provide a count of how many
subdirectories are within a directory, including "." and "..", but not
including files.  Prior to this patch, pvfs2 tried to maintain this, but the
count was a little off, and it couldn't keep up with subdirs added or removed
by other clients anyway.  This led to negative st_nlink counts, with some
possibly odd side effects.

"Making st_nlink track the "right" value would be very expensive in pvfs2
because you might cross meta servers to get the information, and is of
dubious practical use anyway :)   For historical reference, PVFS1 kept
this value fixed as well"


SVN revision 4484