Attached patches implements trusted ports and networks similar to what PVFS1 implemented. It adds a new <Security> </Security> tag to the fs.conf config file and parse utilities accordingly. We can edit these later if there is a strong need.
Within these tags, you have to specify TrustedPorts <Range of ports> and
TrustedNetwork tcp://<network> tcp://<mask>.
I dont know how to support this cleanly on  non tcp/ip bmi
architectures, so currently this only enables it for the bmi_tcp module
alone. (i guess we could do it for bmi_ib module as well but not done
until someone says that is the right thing to do...)

Also adds support for a --trusted argument to the pvfs2-genconfig script
that will prompt for the trusted port ranges and the network address/masks and write it out
to the fs.conf and server.conf file.

Configure now supports a --enable-trusted to enable all these features.
(NOTE that the fs.conf tags (or lack of one) still allow you to not
select these features even if configure was asked to compile with this
option).

this patch also cleans up all the warnings that arises with gcc
(3.4) and --enable-strict build mode.


SVN revision 4649