merges fixes that went into head for consolidating I/O paths (read, write) and (readv, writev)
Also contains the hooks for implementing readx, writex natively in PVFS2. (the vfs patch is part of the patches
sub directory). This involved adding the following infrastructures
a) Have a way of sending a variable length trailer as part of the upcalling mechanism. Currently the readx, writex
implementation upcalls the file offsets, length pairs as this variable length trailer.
b) Client-core has a new upcall request (called FILE_IOX: very unimaginative name :))
which now does a isys_io with file_req set to req_hindexed as opposed to contiguous for FILE_IO upcalls.
c) Client-core does a read of the upcall from the device, if it finds that trailer is present, it does another read
from the device.
d) kmod which maintains the pvfs2_kernel_op_t structures in the request list needs to keep the
structure around in the list (and not insert in the in-progress hash table) if the op has an upcall trailer.
kmod has hooks for feeding either the op or the trailer depending on the which read call comes into the device.
It also has hooks for delaying adding the op to the hash table until the trailer is picked off the device.
e) Added a bunch of test programs for the new readx/writex system calls.. Seems to pass tests on an x86 laptop!

Also contains fixes and cleanups of the readdir/readdirplus implementattion which required a variable length trailer as
part of the downcall. These cleanups enable a very symmetrical implementation of variable length upcall
and downcalls!


SVN revision 5286