When TroveSyncMeta is disabled, databases need to be synced periodically.  Instead of using a timer to perform a sync periodically, we use the new sync coalescing code to allow a sync to occur only in cases where the server is experiencing low load.  Basically, the sync operations are coalesced, but the response from the server is returned immediately.  Once the number of operations queued drops below the low watermark, this signals that the coalesced operations should be synced.  This is controlled via the options in the config file:
With TroveSyncMeta set to 'no', the responses from metadata operations are returned immediately, independent of whether an operation triggers a sync of the databases or not, and the following are the default:

CoalescingHighWatermark infinity
CoalescingLowWatermark  1

With this configuration, we never sync under high-load, and wait till the server idles to sync.  Alternative configurations with TroveSyncMeta set to no might be:

CoalescingHighWatermark 50
CoalescingLowWatermark 0

In this configuration, we sync after every 50 operations.  The sync may occur while the server is still under heavy load servicing metadata operations.  Also, there is no timeout between metadata operations before a sync.


SVN revision 5317