- cleanups to the vfs code for preventing it from retrying or timing out since client-core has much more information than the VFS in determining timeouts etc... Thanks to Phil for all the suggestions and the original idea!
The core idea is as follows
When client-core is about to exit, the kernel closes the device and at that point
the kernel decides to purge all operations that are waiting to be serviced or which are
being currently serviced.
Subsequent to the purge wakeup, all such operations go on a timed sleep.
By default all operations go on a permanent sleep interruptible by a signal of course.
When client-core restarts, there is a possibility of reading operations from the device
before it remounts all the file systems. The kernel now returns only those operations
whose fsids are already mounted.
- Deprecated opening the device in blocking mode.
- Misc. cleanups to the code (moving a lot of the duplicated extern definitions
into the common kernel header file).
- Added a few more gossip debug statements to client and client-core.


SVN revision 5382