[pcarns]: tacl-xattr-homedir.patch: ------------------------- This makes tacl-xattr.sh slightly more portable.  Some Linux distributions have adduser utilities that do not create the home directory for you.  This patch explicitly does a mkdir -p and chmod after adduser to make sure that required home directories exist.
tacl-xattr-symlink.patch:
-------------------------
This is an important fix to the test script.  It was using the -L argument to getfattr to traverse symbolic links when dumping the ACLs from the test directory.
This
leads to unpredictable results because there was no way to tell whether getfattr would traverse a real directory or its symbolic link first (it depends on the
dirent
order) and the other would always be left out.  ACLs are not supported on symbolic links anyway, so it was just adding noise to the test script.

xattr-symlink.patch:
-------------------------
This is the only change to PVFS2 itself.  Like most Linux file systems, PVFS2 does not support xattrs on symbolic links (despite what is implied by man pages).
This
is due to the fact that symbolic links have 777 permissions by default that would allow anyone on the system to store xattrs in any symbolic link.  This patch
updates
PVFS2 semantics slightly, however, to be more in line with how other file systems implement this.  In particular listxattr() is now allowed (it just returns that
there are zero entries), and setxattr() is implemented to return EPERM rather than EOPNOTSUPP.

pread-pwrite.patch:
-------------------
This fixes a bug in a patch that I submitted earlier to provide a simple alternate AIO implementation.  It defines _GNU_SOURCE in a limited area for dbpf so that
we
can get proper definitions of pread() and pwrite() on Linux.  I tried using _XOPEN_SOURCE=500, but it will break any .c file that includes dbpf.h due to
incompatibilities with Berkeley DB.

zero-dfile.patch:
-------------------
This fixes a bug in the getattr handling on pvfs2-server if it happens to find an attribute structure with the dfile array zeroed out.  In this case, it needs to
set
the attr flag appropriately to prevent the response encoder from segfaulting while processing the array in the response structure.  This condition is very hard to
trigger, but the server should be able to gracefully report the error rather than crashing.

bind-specific.patch:
--------------------
This patch adds a new config file option (TCPBindSpecific) that if enabled tells the server to bind only to its specific IP address rather than using INADDR_ANY.
This is particularly helpful in failover scenarios where you would like one physical machine to assume two ip addresses (and two servers) to run simultaneously
when
another server crashes.  Without this patch you would need to select a different port on each server to prevent collisions.


SVN revision 5634