a) Use calloc() instead of malloc() for bmi-tcp's memalloc() callback b) memset() the events structure used by epoll_ctl Both the above help prevent valgrind from spewing out uninitialized bytes warnings..
Rework the config file integrity checks into the fs-add state machine based on Sam and RobL's suggestions.
The reason we do this is so that we keep the status quo of hiding config files
from the system interface users. isys_fs_add continues to be the system interface
whose state machine internally verifies that all servers are using the same fs.conf
file by checking their file sizes and their cryptographic check sums and see if they match.

pvfs2-ping.c goes back to its original form without the call to PINT_fetch_config_list()
which is now deprecated and unused (it is still left in case we find a use for it someplace
else).

Also fixes the breakages caused by the recent export options merge in the test sub directory
and a few gcc 4 warnings with void* arithmetic..


SVN revision 5639