Merge posix-extensions-branch to HEAD This branch implemented patches to the 2.6.16 kernel for the proposed POSIX I/O extensions and those patches are under the patches subdirectory. It also implements the PVFS2 specific hooks for these system calls.
Tools that may be of immediate use to the pvfs2 general audience is the pvfs2-lsplus utility in src/apps/admin
that should be noticeably faster than the pvfs2-ls utility if there are a lot of objects in a single directory.
Other features are left out by configure and are not even built if the kernels do not support those features
and callbacks.


SVN revision 5656