[on behalf of Phil]:
pvfs2-change-fsid.patch
-----------------------
This patch adds a new admin utility called pvfs2-change-fsid.  It can be used to modify the fsid on an existing file system.  This is helpful if you duplicate a file system (from backup, clone, etc.) and want to make sure that the new one does not conflict with the old one.


SVN revision 6438