[on behalf of Phil]:
pvfs2-check-server-gen-mntent.patch
-----------------------------------
This patch does two things.  First, it adds a utilty function called PVFS_util_gen_mntent() that can be used to create a mntent structure in memory without parsing a tabfile.  Previously there wasn't any clean way to construct this on the fly.  Secondly, it adds a utility (using the gen_mntent() function) called pvfs2-check-server.  pvfs2-check-server can be used to probe a specific pvfs2 server to see if it is running and responding to protocol requests.  Since it does not rely on a tab file, it can talk directly to one specific server without having to retrieve configuration information from the primary server listed in the tab file.  This utility is useful for watchdog purposes when you want to find out if one specific server is working properly.


SVN revision 6439