[on behalf of Phil]:
The documentation directory in the source code already contains a document describing how to set up PVFS for high availability using version 1 of heartbeat.  This patch adds a similar document describing how to do the same thing with version 2, which offers several advantages.  It also adds a new examples directory with a few scripts that correspond to the example in the document.  This could probably be made more generic, but in its current form it may at least save someone some time in setting up an HA environment with a newer version of heartbeat.  The architecture and syntax is now quite a bit different.


SVN revision 6440