[on behalf of Phil]:
This patch adds a new configuration parameter called "RootSquashExceptions".  It has the same value syntax as the existing RootSquash parameter, but it allows you to list hosts or subnets that are exempt from root squashing.

This is helpful if you want to root squash all clients (or at least a large set of them) but still allow root access from particular admin or power user clients.  That was possible before, but probably tedious unless the node you wanted to exclude just happened to fall in a convenient subnet.


SVN revision 6444