[on behalf of Phil]:
pvfs2-aio-cancel.patch
----------------------
This patch fixes a bug in the I/O cleanup path on the server side.  In cases where a flow needed to cancel pending I/O operations, the trove cancel function was calling aio_cancel() directly.  This doesn't work correctly if the alt-aio implementation is used.

pvfs2-root-squash-address.patch
-------------------------------
This fixes a bug in the root squash checking on the server side.  The routine that compares a client address against the root squash list was using getsockname() rather than getpeername().  The former retrieves the server's address rather than the client's.

pvfs2-ls-rm.patch
-----------------
This is an interim fix for the concurrent "rm -rf" and "ls" problem that was recently discussed on the mailing list.  It sounds like the long term direction is to switch to using entry names as dirent tokens, but this patch fixes the majority of cases in the mean time without a protocol change.  The problem in the case I was seeing was a cache conflict between the two clients (the ls was caching tokens in the pcache that caused rm to get the wrong position).  The token is 64 bits wide, but only the first 32 bits are used (the START and END values are near the top of the 32 bit range).  This patch takes advantage of the extra top 32 bits on the server side to set a unique identifier in the token for each "readdir session" so that their cache entries do not collide.  The client is not aware of this change because it treats the token as an opaque value.  A readdir session begins when a client requests the START position.

pvfs2-client-buffer-logging.patch
---------------------------------
I don't know if there is any interest in this, but this adds some debugging to the buffers used in the kernel module.  On startup, pvfs2-client will print the buffer pointers (whether debugging is enabled or not).  There are also new debugging messages that will show the first byte of each memory buffer passing through the kernel if enabled.  These logging messages were added to help track down what ended up being a server side problem (see pvfs2-aio-cancel.patch), but we kept it in case it is useful in the future.


SVN revision 6447