Remove "fast path" initial lookup state.  Per comments from Phil, the ncache will never hit anything but a single element, and this full-path query into the ncache is historical.
In the case that lookup is called on a single element, the first iteration
through the rest of the lookup states will find it anyway.  That may cost
a few more states, but no extra server queries.  Removing this code, however,
should make things easier to understand in the future.


SVN revision 6523