adding a special seek operation for pvfs2 directories.  The default vfs seek doesn't work because we have special meanings for f_pos on directories:
- 0 and 1: fake "." and ".." entries
- PVFS_READDIR_END and PVFS_READDIR_START: special names
- all else: pvfs readdir token value

Who would want to seek on a directory?  Well, glibc occasionally decides to
for some reason: glibc-*/sysdeps/unix/sysv/linux/getdents.c.  There is still
a flaw in that we can't seek to offset 3 or 4 if someone happens to do that
(see code comments).  See this trac entry for how to trigger (and lose dir
entries) on some systems:
https://trac.mcs.anl.gov/projects/pvfs/ticket/11


SVN revision 6565