more dcache changes: what's the worst that could happen?  :-)
Toning down the d_delete function so that it only throws away entries that
have a bad inode.  Updated the revalidate path to mark inodes bad when a
revalidation fails.  This should work fine for the simul test case, and also
leaves inodes alone if we don't specifically know of anything wrong with them.

Our NFS re-exporting ability right now is dependent on good inodes not
disappearing out from under knfsd.  knfsd drops file references and expects
to be able to re-get them later based on inode number.  The inode number is
only 32 bits and not sufficient for PVFS to reliably locate the file handle.
The normal solution is to implement vfs export ops (which allow you
to use identifiers more sophisticated than inode number), but this
would be really hard because of difficulty in finding directory parents
in PVFS.


SVN revision 6583