gm-mx-autoconf
Convert GM and MX to use autoconf tests to find includes and libs.  Move
these tests into new config files to keep the top-level Makefile clean.
This fixes a configuration problem with GM on x86_64 where libs are
in lib64/, and generalizes both GM and MX to be a bit more clever at
autodetection.

Signed-off-by: Pete Wyckoff <pw@osc.edu>
---
 Makefile.in                    |   32 +++-----
 configure                      |  167 ++++++++++++++++++++++++++++++++--------
 configure.in                   |   49 ++----------
 maint/config/gm.m4             |   72 +++++++++++++++++
 maint/config/mx.m4             |   73 +++++++++++++++++
 src/apps/admin/pvfs2-config.in |    8 +-
 src/io/bmi/bmi_gm/module.mk.in |   45 ++++++++----
 src/io/bmi/bmi_mx/module.mk.in |   40 +++++++---
 8 files changed, 363 insertions(+), 123 deletions(-)
 create mode 100644 maint/config/gm.m4
 create mode 100644 maint/config/mx.m4


SVN revision 6626