threaded kmod helper
Add --enable-threaded-kmod-helper option to build a threaded
pvfs2-client-core, that is, an executable linked against
pvfs2-threaded.so.  Take care to make sure only what libraries
are needed based on --enable-static, --enable-shared, and this
new option.

Remove the "--threaded" argument from pvfs2-client.  Whatever version
of client-core that was configured is all that is available at runtime.
It will always be installed as pvfs2-client-core.


SVN revision 6638