shared lib deps
For proper symbol versioning, at shared library build time, it is necessary
to specify all the other shared libraries that will be required to run
the one we are about to create.  The particular place this shows up as a
problem is for libibverbs.so when building bmi ib.  If you don't tell the
linker that you'll need libibverbs.so later, at runtime, it will pickup
the 1.0 rather than the 1.1 symbols.


SVN revision 6792