library versioning
This patch adds a version code to the installed shared libraries,
libpvfs2.so and possibly libpvfs2-threaded.so.  Library versioning
is considered a good thing, in that apps can know what version of the
library they need.  Use the actual PVFS version numbers to build
the library version string.  This is not always the right thing to do,
but works for us because we pay attention to the meaning of "major",
"minor" and "release" changes.  Adapted from a patch by Justin Bronder.


SVN revision 6807