fix at least part of bug reported by Tony Kew: http://www.beowulf-underground.org/pipermail/pvfs2-users/2008-June/002476.html
The "basic_dist" distribution did not work on file systems with more than
one server because of the number of datafiles being set higher than what is
supported by the distribution (1 in this case).  Reorganized so that the
logic for setting number of datafiles is in one spot and is
performed only after the final distribution is set.


SVN revision 7180