Partial fix to a subtle sockio bug: don't loop over socket peek and sum results, because if there is only 1 byte in the socket (for example), then it will add 1 every time and eventually add up to the bmi header size.  This causes bmi to try to read a complete header out of the socket.
This made it possible to hang a server by telnet'ing to port 3334,
typing one character, and hitting return.  BMI would get stuck trying to
receive 24 bytes in an nbrecv() call that seems to block indefinitely (maybe a
secondary bug?).

This is only a partial fix, because now rather than a hang we get a high cpu
load until the socket is disconnected.


SVN revision 7671