Fixup static and inline functions.
Change all static inline or inline static functions in .c files to simply
static.

Real inline functions should be declared and defined static inline in a
.h file.

Reason is that the inline keyword makes no difference to the optimizer
when it decides whether to inline or not. The inline keyword tells the
compiler not to warn on unused functions from header files and the
linker not to warn on multiple definitions.

Declaring a function only inline doesn't mean anything useful in since
the compiler won't be able to look into a different .o file to get the
code to inline. They must be declared and defined static inline in a .h
file so the compiler will have the code to inline.

SVN revision 11663