Kmod: refactor old-style writev function
The client core sends responses via a series of io_vecs... there
are two versions of the function that processes the io_vecs,
an old-style one that uses copy_from_user and a newer one that
uses copy_from_iter.

This commit seriously cleans up the old-style function and should
reduce the number of times the downcall buffer is copied around in
kernel memory by 50%.

Also... Back in 2006, in commit r5975, pvfs2_devreq_aio_write was added
as a wrapper to pvfs2_devreq_writev. 
  
pvfs2_devreq_aio_write = file_operations.aio_write
  
   pvfs2_devreq_writev = file_operations.writev
  
Nobody past 2.6.19 needs .writev, everyone passes through the
wrapper until the whole shebang changes with the addition of
.write_iter around 4.0.

Since "nobody" uses .writev, and there have been some changes
in the arguments to .aio_write, the wrapper is more trouble than
it is worth. This gets rid of the wrapper.



SVN revision 12578