Allow PVFS2_VFS_OP_FEATURES upcall even if the filesystem is not yet (re)mounted.  Without this, the filesystem becomes unresponsive after the client core is restarted.
The client core does not consider the remount done until after the
PVFS2_DEV_REMOUNT_ALL ioctl returns.  The kernel makes the features
request before the ioctl returns, so the client core must actually
respond to it.

The client core ignores requests made before the remount is finished.
Perhaps this should be revisited as it can cause the kernel to hang
waiting on a response that will never come.

This is only half of the bugfix.  The other half is in the upstream
kernel.  The out of tree kernel does not ever make the features request.


SVN revision 12793