Check whether remote_map is set before processing an unexpected receive.
When testing before, logging was on and was hiding a race condition
between the client and server. With debugging turned off, the race
condition would occassionally crash the server. After the server accepts
a connection request from a client, both the client and server wait for
RDMA_CM_EVENT_ESTABLISHED. The server then registers the method address
for the new connection with BMI, and the client can post its first send.
While all that is going on, the job layer is looping in the
bmi_thread_function(), calling BMI_testunexpected(), which checks the
receive queue. The problem is that the client can potentially see that
the connection is established and post a send before the server finishes
registering the new address. Since the server has to post receives
before accepting the new connection, the client's send successfully
makes it to the receive queue and is found by testunexpected(). At this
point, it's a race for the accept/listener thread to register the method
address for that connection before testunexpected() tries to retrieve
that address when processing the receive. To handle this, the server now
checks that the connection's remote_map has been set before trying
to process a receive in BMI_rdma_testunexpected().



SVN revision 12937