Fixed a problem that was preventing the server from being able to be shut down gracefully.
With rdma_get_cm_event() in its default, blocking mode, it would block 
indefinitely while waiting for a new event. This meant that when you
tried to stop the server, the listener thread would never see that it
needed to shutdown unless a client tried to communicate again. So the
server would either wait indefinitely or it would try to accept a new
connection at the same time it was shutting down, which would cause a
seg fault. Changing rdma_get_cm_event() to non-blocking solved this
issue, but that meant the server was in a busy loop nearly 100% of the
time, so I added a poll() on the listener's event channel before calling
rdma_get_cm_event(). That way it can timeout and check for shutdown.
Since I added the poll(), I brought back the timeout that was removed
in r12939, but I still need to experiment to find the best value.


SVN revision 12940