Moved the rdma_conn info into the rdma_connection_t structure.
Needed to rethink the way we use the rdma_conn structure and the
information that is stored in it, since that information needs to hang
around longer than when we were using TCP to initiate IB connections.
This also required some changes to the way we allocate new connections,
but, overall, I think it makes things a bit cleaner. Note, the peername
field is mostly used for debug statements, so it could possibly be
removed altogether in the future.



SVN revision 13026