Close client connections before destroying the listening id and event channel on the server to prevent "Bad file descriptor" errors.
This seems obvious, but under normal conditions, a client connection is
closed on the server when it receives a MSG_BYE from a client, which
happens before BMI_rdma_finalize() is called. Therefore, it didn't
become apparent that things were out of order here until I began testing
failover. If the server goes down while a client is still connected, it
will never receive a BYE message, which means when it got here it was 
destroying the listening event channel, and then it was trying to
destroy ids of client connections that were associated with that
channel. Everything was still working, but it would spit out "Bad file
descriptor" errors. This is fixed now.


SVN revision 13041