server: fix off-by-one ACL-length error...
In the last few years we have tried to get all the code to agree
on whether or not xattrs should be treated as text or data. Are
xattrs NULL terminated, or not? Do we count the NULL in the xattr
length, or not?

Blobs of ACL data are stored in xattrs on files and directories. 
All I can figure is that we overlooked the ACL code as we adjusted other
xattr code, leading to a situation where setting an ACL on a file or
directory caused the file or directory to become broken. Even a
lookup would get EINVAL by everyone but root - if it's root we don't
bother checking permissions and ACLs and other plebeian stuff
like that <g>.

The problem boiled down to our code that verifies the ACL data in an
xattr. Math is done on the ACL data in an xattr to ensure that it
contains exactly enough bytes to represent one acl header and N acl entries. 
With our off-by-one error, our verification code would decide all
ACLs were bad, resulting in EINVAL.

The server has to support ACLs, I think only the kernel module uses
ACLs. The kernel module could put a valid ACL on an object with setfacl(1),
but after that lookup service_operations on the object would get
EINVAL back from the server.



SVN revision 13084