server: remove unused ifdef and associated code, PVFS_USE_OLD_ACL_FORMAT
Apparently, before the formal posix way to store ACL data
in extended attributes was settled, there was a slightly different
format in use.

We had (vestigial?) support for the old method, in the form of
ifdefs. Code like this was sprinkled around in our source files where ACL
stuff happens:

#ifdef PVFS_USE_OLD_ACL_FORMAT
  do it the old way
#else
  do it the posix way
#endif

It seems we don't support any Linux flavors old enough to use
the old format. And there's no reasonable way to set PVFS_USE_OLD_ACL_FORMAT
short of manually adding it to CFLAGS at configure time. And it makes
the code hard to read.

So I got rid of it.



SVN revision 13086