Makes the NIC/HCA port used for IB/RoCE communication configurable.
On the server, you configure it via the config file. There's a new
BMIOpts option you can use to specify BMI-specific options that will be
passed to BMI when it is initialized. Configure the port by adding
"ib_port=N" to BMIOpts, where N is the port to use.

On the client, there are now two ways to pass options to BMI. If you are
not using the client-core, you can add "bmi_opts=NNN" as an option in
the tab file. Otherwise, you can use the --bmi-opts=NNN command-line
option when starting the client-core. In both cases, NNN is a comma-
separated list of options. You can tell BMI to use a specific port by
including "ib_port=N" in those options.



SVN revision 13137