Fixes connection-based reference counting to handle flush errors.
The previous implementation of reference counting wasn't preventing
things from getting shut down/destroyed when there were still
outstanding requests on a queue pair. Depending on the environment
it was running in, this would either cause it to segfault after
receiving a flush error because it was trying to access the
corresponding connection structure, which was being freed too early,
or to never see any flush errors at all.


SVN revision 13174