Fixed up the client-core enough that it can send sid_count and sid_array back to the kernel module in the mount response downcall trailer_buf.
Fixed up a version of the upstream kernel module (starting with commit
74e938c2) to fetch the sid_count and sid_array out of the downcall,
and send them back to the client-core in the subsequent getattr
upcall trailer_buf. The ability to process upcall trailer-bufs 
was achieved by resurrecting the op_linger strategy into the upstream
kernel module.

Further fixed up the client-core so that it could fetch the sid_count and
sid_array from the getattr upcall trailer_buf.

Moved ntime to the bottom of PVFS_sys_attr_s because it breaks
the protocol where it is, but this alone isn't a "real" fix.



SVN revision 13200