bmi_rdma: Only increment refcnt when an RDMA write is completely done
This fixes a refcnt issue when doing large reads from multiple servers.
When a client does a large read, what actually happens is an RDMA write
from the servers to the client. The maximum size of a single RDMA write
is 256K (corresponding to the default flow buffer size), and the
default block/strip size for OFS is 64K. So when using multiple
servers, a single RDMA write is broken down into up to four 64K chunks.
We are incrementing the refcnt for each of these chunks, rather than
incrementing once after all chunks have been written and the entire
RDMA write operation is completed. This means if the block size of the
read is greater than 64K * the number of servers we can end up with up
to three extra references per RDMA write. Since we decrement the refcnt
when we get a completion for an outstanding operation, and we only get
a completion for the full RDMA write, not the individual chunks, these
extra references never get cleaned up and prevent the connection from
being closed at cleanup time. This commit fixes that by only
incrementing the refcnt after the entire RDMA write has completed, when
all chunks have been written.