[Bug fix] bmi_rdma: Fix issue causing servers to hang on shutdown
In BMI_rdma_finalize(), allow OrangeFS servers to send BYE messages and
disconnect queue pairs when those servers are acting as clients to other
servers. Previously, only OrangeFS clients would do this, and servers
would wait until the BYE was received to tear down the connection. We
would check for a valid listen_id to determine whether it was a client
or a server. When it was written that way, servers never acted as
clients, but that is not the case anymore. Now, when an OrangeFS server
connects to another server, it actually results in two connections: one
as a server and one as a client. But even when acting as a client, an
OrangeFS server will always have a valid listen_id. In those cases, we
would see there was a valid listen_id, assume it was a server, skip the
steps of sending BYEs and disconnecting, and wait to receive a BYE from
the other side. However, since the other side is also a server and
doing the same thing, that BYE will never come. This issue was masked in
the original IB code because even if there were open connections at
shutdown, we would close them and shut down anyway, ignoring any
outstanding requests. But with librdmacm, those outstanding requests
caused segfaults, so changes were made to reference counting to delay
shutdown until all outstanding requests have been flushed (see commit
382a5bc18b0f7d415c68b36f21d507a142fa3cce). After that change, when
shutting down servers, we now get stuck in an infinite loop waiting for
requests to be flushed, since rdma_disconnect() needs to be called to
trigger those flush errors and the server doesn't call rdma_disconnect()
until it receives a BYE message that will never come in this case. This
problem is solved by removing the check for a valid listen_id and
allowing both clients and servers to send BYE messages and call
rdma_disconnect() in BMI_rdma_finalize().